limit calls to MatchFactory reset in tests; add no retry option to HootTest
