+++ docs/developer/HootenannyCodeBestPractices.asciidoc
@@ -28,6 +28,7 @@ getter methods.
 logic needed to be executed by the command.
 * Make use of configuration options via the `Configurable` interface to keep functionality adapatable 
 to a broad range of data inputs.
+* Avoid setting the configuration in a global manner with `conf().set(...)` when possible.
 * Favor the visitor pattern via classes ineriting `ElementVisitor` for map data processing. Use 
 `OsmMapOperation` only when the data processor needs a copy of a complete map to performs its logic.
 * Use classes inheriting `ElementCriterion` to identify data based on its properties.
