+++ hoot-core-test/src/test/cpp/hoot/core/algorithms/WayJoinerTest.cpp
@@ -28,17 +28,13 @@
 //  Hoot
 #include <hoot/core/TestUtils.h>
 #include <hoot/core/algorithms/NonIntersectionWayJoiner.h>
-#include <hoot/core/algorithms/WayJoinerAdvanced.h>
 #include <hoot/core/algorithms/WayJoinerBasic.h>
 #include <hoot/core/algorithms/splitter/HighwayCornerSplitter.h>
 #include <hoot/core/algorithms/splitter/IntersectionSplitter.h>
-#include <hoot/core/conflate/UnifyingConflator.h>
 #include <hoot/core/io/OsmXmlReader.h>
 #include <hoot/core/io/OsmXmlWriter.h>
-#include <hoot/core/ops/OpExecutor.h>
 #include <hoot/core/util/ConfigOptions.h>
 #include <hoot/core/elements/MapProjector.h>
-#include <hoot/core/visitors/RemoveTagsVisitor.h>
 
 namespace hoot
 {
@@ -49,15 +45,13 @@ class WayJoinerTest : public HootTestFixture
   CPPUNIT_TEST(runSimpleTest);
   CPPUNIT_TEST(runCornerSplitterTest);
   CPPUNIT_TEST(runIntersectionSplitTest);
-  CPPUNIT_TEST(runConflateTest);
-  CPPUNIT_TEST(runAdvancedConflateTest);
   CPPUNIT_TEST(runNonIntersectionJoinerTest);
   CPPUNIT_TEST_SUITE_END();
+
 public:
 
-  WayJoinerTest()
-    : HootTestFixture("test-files/algorithms/WayJoinerTest/",
-                      "test-output/algorithms/WayJoinerTest/")
+  WayJoinerTest() :
+  HootTestFixture("test-files/algorithms/WayJoinerTest/", "test-output/algorithms/WayJoinerTest/")
   {
     setResetType(ResetBasic);
   }
@@ -119,69 +113,6 @@ public:
                      _inputPath + "WayJoinerIntersectionSplitterExpected.osm");
   }
 
-  void runConflateTest()
-  {
-    OsmXmlReader reader;
-    OsmMapPtr map = std::make_shared<OsmMap>();
-    reader.setDefaultStatus(Status::Unknown1);
-    reader.read("test-files/ToyTestA.osm", map);
-    reader.setDefaultStatus(Status::Unknown2);
-    reader.read("test-files/ToyTestB.osm", map);
-
-    OpExecutor(ConfigOptions().getConflatePreOps()).apply(map);
-
-    UnifyingConflator conflator;
-    conflator.apply(map);
-
-    OpExecutor(ConfigOptions().getConflatePostOps()).apply(map);
-    MapProjector::projectToWgs84(map);
-
-    RemoveTagsVisitor hashRemover(QStringList(MetadataTags::HootHash()));
-    map->visitRw(hashRemover);
-
-    OsmXmlWriter writer;
-    writer.setIncludeCompatibilityTags(true);
-    writer.setIncludeHootInfo(true);
-    writer.setIncludePid(true);
-    writer.write(map, _outputPath + "WayJoinerConflateOutput.osm");
-
-    HOOT_FILE_EQUALS(_outputPath + "WayJoinerConflateOutput.osm",
-                     _inputPath + "WayJoinerConflateExpected.osm");
-  }
-
-  void runAdvancedConflateTest()
-  {
-    OsmXmlReader reader;
-    OsmMapPtr map = std::make_shared<OsmMap>();
-    reader.setDefaultStatus(Status::Unknown1);
-    reader.read("test-files/ToyTestA.osm", map);
-    reader.setDefaultStatus(Status::Unknown2);
-    reader.read("test-files/ToyTestB.osm", map);
-
-    OpExecutor(ConfigOptions().getConflatePreOps()).apply(map);
-
-    UnifyingConflator conflator;
-    Settings conf;
-    //  Use the Advanced way joiner
-    conf.set(ConfigOptions::getWayJoinerKey(), WayJoinerAdvanced::className());
-    conflator.apply(map);
-
-    OpExecutor(ConfigOptions().getConflatePostOps()).apply(map);
-    MapProjector::projectToWgs84(map);
-
-    RemoveTagsVisitor hashRemover(QStringList(MetadataTags::HootHash()));
-    map->visitRw(hashRemover);
-
-    OsmXmlWriter writer;
-    writer.setIncludeCompatibilityTags(true);
-    writer.setIncludeHootInfo(true);
-    writer.setIncludePid(true);
-    writer.write(map, _outputPath + "WayJoinerAdvancedConflateOutput.osm");
-
-    HOOT_FILE_EQUALS(_outputPath + "WayJoinerAdvancedConflateOutput.osm",
-                     _inputPath + "WayJoinerAdvancedConflateExpected.osm");
-  }
-
   void runNonIntersectionJoinerTest()
   {
     OsmXmlReader reader;
@@ -198,9 +129,7 @@ public:
 
     HOOT_FILE_EQUALS(_outputPath + "NonIntersectionWayJoinerOutput.osm",
                      _inputPath + "NonIntersectionWayJoinerExpected.osm");
-
   }
-
 };
 
 CPPUNIT_TEST_SUITE_NAMED_REGISTRATION(WayJoinerTest, "slow");
