+++ hoot-core-test/src/test/cpp/hoot/core/conflate/CookieCutterTest.cpp
@@ -61,7 +61,7 @@ public:
     : HootTestFixture("test-files/conflate/CookieCutterTest/",
                       "test-output/conflate/CookieCutterTest/")
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   void runTest()
