+++ hoot-core-test/src/test/cpp/hoot/core/conflate/matching/OptionsValidatorTest.cpp
@@ -63,7 +63,7 @@ public:
 
   OptionsValidatorTest()
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   void runValidateSizeUnequalTest()
