+++ hoot-core-test/src/test/cpp/hoot/core/conflate/network/EdgeStringTest.cpp
@@ -57,7 +57,7 @@ public:
 
   EdgeStringTest()
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   void basicTest()
