+++ hoot-core-test/src/test/cpp/hoot/core/conflate/network/EdgeSublineMatchTest.cpp
@@ -43,7 +43,7 @@ public:
 
   EdgeSublineMatchTest()
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   void basicTest()
