+++ hoot-core-test/src/test/cpp/hoot/core/conflate/network/NetworkDetailsTest.cpp
@@ -59,7 +59,7 @@ public:
 
   NetworkDetailsTest()
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   NetworkDetailsPtr createSampleDetails()
