+++ hoot-core-test/src/test/cpp/hoot/core/conflate/poi-polygon/PoiPolygonMatchCreatorTest.cpp
@@ -53,7 +53,7 @@ public:
 
   PoiPolygonMatchCreatorTest()
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   void basicTest()
