+++ hoot-core-test/src/test/cpp/hoot/core/elements/MapProjectorTest.cpp
@@ -68,7 +68,7 @@ public:
 
   MapProjectorTest()
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   virtual void setUp()
