+++ hoot-core-test/src/test/cpp/hoot/core/ops/DuplicateNodeRemoverTest.cpp
@@ -57,7 +57,7 @@ public:
 
   DuplicateNodeRemoverTest() : HootTestFixture("test-files/", UNUSED_PATH)
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   void runBasicTest()
