+++ hoot-core-test/src/test/cpp/hoot/core/ops/FindIntersectionsOpTest.cpp
@@ -49,7 +49,7 @@ public:
   FindIntersectionsOpTest() :
   HootTestFixture("test-files/ops/FindIntersectionsOp/", "test-output/ops/FindIntersectionsOp/")
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   void runToyTest()
