+++ hoot-core-test/src/test/cpp/hoot/core/ops/ReprojectToPlanarOpTest.cpp
@@ -60,7 +60,7 @@ public:
   ReprojectToPlanarOpTest() :
   HootTestFixture("test-files/ops/ReprojectToPlanarOp/", "test-output/ops/ReprojectToPlanarOp/")
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   void runTest()
