+++ hoot-core-test/src/test/cpp/hoot/core/ops/WaySplitterOpTest.cpp
@@ -57,7 +57,7 @@ public:
   WaySplitterOpTest() :
   HootTestFixture("test-files/ops/WaySplitterOp/", "test-output/ops/WaySplitterOp/")
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   void runTest()
