+++ hoot-core-test/src/test/cpp/hoot/core/test/TestSetup.cpp
@@ -39,7 +39,6 @@ _confs(confs)
 
 TestSetup::~TestSetup()
 {
-  reset();
 }
 
 void TestSetup::reset()
