+++ hoot-rnd/src/test/cpp/hoot/rnd/conflate/multiary/MultiaryPoiGenericTest.cpp
@@ -50,7 +50,7 @@ public:
 
   MultiaryPoiGenericTest()
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   /**
