+++ hoot-rnd/src/test/cpp/hoot/rnd/conflate/multiary/MultiaryPoiMergeCacheTest.cpp
@@ -51,7 +51,7 @@ public:
 
   MultiaryPoiMergeCacheTest()
   {
-    setResetType(ResetAll);
+    setResetType(ResetAllNoMatchFactory);
   }
 
   /**
