+++ test-files/algorithms/WayJoinerConflateTest/WayJoinerAdvancedConflateExpected.osm
