+++ test-files/algorithms/WayJoinerConflateTest/WayJoinerConflateExpected.osm
